bug fix in synth code

+++ synthetic_evaluation.py
@@ -39,7 +39,7 @@ def main():
                         help="Number of synthetic functions used for the evaluation. Must be an integer value.")
     parser.add_argument("--nr-repetitions", type=int, default=5, required=True,
                         help="Number of repetitions for each measurement point. Must be an integer value.")
-    parser.add_argument("--noise", type=int, default=1, required=True,
+    parser.add_argument("--noise", type=float, default=1, required=True,
                         help="Percentage of induced noise. Must be an integer value.")
     modeling_options = parser.add_argument_group("Modeling options")
     modeling_options.add_argument("--median", action="store_true", dest="median",
